Add more tests for related issues
